[BP-1] - renamed Profile to Account
