+++ src/main/kotlin/com/bp/app/account/AccountRepository.kt
@@ -1,8 +1,8 @@
-package com.bp.app.profile
+package com.bp.app.account
 
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface ProfileRepository : JpaRepository<Profile, Long> {
+interface AccountRepository : JpaRepository<Account, Long> {
 }
